+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* add REST route to return `plugins_api()` like data for installed plugin
 
 #### 1.3.6 / 2022-03-02
 * update Freemius/wordpress-sdk